# docs/swagger.yaml
definitions:
...
  domain.Request:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      rejectNotes:
        type: string
      status:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
      verifierID:
        type: integer
    type: object
...
  domain.VolunteerDetails:
    properties:
      createdAt:
        type: string
      departmentID:
        type: integer
      id:
        type: integer
      status:
        type: integer
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  dto.AddRejectNoteRequest:
    properties:
      notes:
        type: string
    type: object
  dto.ApplicantCreateDTO:
    properties:
      email:
        type: string
      name:
        type: string
      surname:
        type: string
    required:
    - email
    - name
    - surname
    type: object
  dto.ApplicantResponseDTO:
    properties:
      country_id:
        type: integer
      department_id:
        type: integer
      dob:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      mobile:
        type: string
      name:
        type: string
      resident_country_id:
        type: integer
      role_id:
        type: integer
      surname:
        type: string
    type: object
  dto.AppplicantUpdateDTO:
    properties:
      country_id:
        type: integer
      department_id:
        type: integer
      dob:
        type: string
      email:
        type: string
      gender:
        type: string
      mobile:
        type: string
      name:
        type: string
      resident_country_id:
        type: integer
      role_id:
        type: integer
      surname:
        type: string
    type: object
...
  dto.ListRequest:
    properties:
      requests:
        items:
          $ref: '#/definitions/domain.Request'
        type: array
    type: object
...
  dto.RequestResponse:
    properties:
      create_at:
        type: string
      id:
        type: integer
      reject_notes:
        type: string
      status:
        type: integer
      type:
        type: string
      update_at:
        type: string
      user_id:
        type: integer
      verifier_id:
        type: integer
    type: object
...
  dto.VolunteerCreateDTO:
    properties:
      department_id:
        type: integer
      status:
        type: integer
      user_id:
        type: integer
    required:
    - department_id
    - status
    - user_id
    type: object
  dto.VolunteerUpdateDTO:
    properties:
      department_id:
        type: integer
      status:
        type: integer
    type: object
info:
  contact: {}
  description: This is a volunteer service API
  title: Onboarding and Volunteer Service API
  version: "1.0"
paths:
  /api/v1/admin/add-reject-notes/{id}:
    post:
      description: Add reject notes
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add Reject Note Request
        in: body
        name: notes
        required: true
        schema:
          $ref: '#/definitions/dto.AddRejectNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Add reject notes
      tags:
      - admin
  /api/v1/admin/approve-request/{id}:
    post:
      description: Approve request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Approve request
      tags:
      - admin
  /api/v1/admin/delete-request/{id}:
    delete:
      description: Delete request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Delete request
      tags:
      - admin
  /api/v1/admin/list-pending-request:
    get:
      description: Get list pending request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListRequest'
      security:
      - bearerToken: []
      summary: Get list pending request
      tags:
      - admin
  /api/v1/admin/list-request:
    get:
      description: Get list request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListRequest'
      security:
      - bearerToken: []
      summary: Get list request
      tags:
      - admin
  /api/v1/admin/pending-request/{id}:
    get:
      description: Get pending request by ID
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RequestResponse'
      security:
      - bearerToken: []
      summary: Get pending request by ID
      tags:
      - admin
  /api/v1/admin/reject-request/{id}:
    post:
      description: Reject request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Reject request
      tags:
      - admin
  /api/v1/admin/request/{id}:
    get:
      description: Get request by ID
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RequestResponse'
      security:
      - bearerToken: []
      summary: Get request by ID
      tags:
      - admin
...
  /api/v1/applicant/:
    post:
      description: Create applicant
      parameters:
      - description: Create Applicant Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApplicantCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Applicant created successfully
          schema:
            type: string
      summary: Create applicant
      tags:
      - applicant
  /api/v1/applicant/{id}:
    delete:
      description: Delete applicant
      parameters:
      - description: Applicant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete applicant
      tags:
      - applicant
    get:
      description: Find applicant by ID
      parameters:
      - description: Applicant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApplicantResponseDTO'
      summary: Find applicant by ID
      tags:
      - applicant
    put:
      description: Update applicant
      parameters:
      - description: Applicant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Applicant Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AppplicantUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Applicant updated successfully
          schema:
            type: string
      summary: Update applicant
      tags:
      - applicant
...
  /api/v1/volunteer/:
    get:
      description: Get all volunteers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VolunteerDetails'
      summary: Get all volunteers
      tags:
      - volunteer
    post:
      description: Create volunteer
      parameters:
      - description: Create Volunteer Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VolunteerCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Volunteer created successfully
          schema:
            type: string
      summary: Create volunteer
      tags:
      - volunteer
  /api/v1/volunteer/{id}:
    delete:
      description: Delete volunteer
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Volunteer deleted successfully
          schema:
            type: string
      summary: Delete volunteer
      tags:
      - volunteer
    get:
      description: Find volunteer by ID
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VolunteerDetails'
      summary: Find volunteer by ID
      tags:
      - volunteer
    put:
      description: Update volunteer
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Volunteer Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VolunteerUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Volunteer updated successfully
          schema:
            type: string
      summary: Update volunteer
      tags:
      - volunteer
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"